requirements.txt
Django==5.2.8
djangorestframework==3.16.1
drf-yasg==1.21.11
django-environ==0.12.0
django-ratelimit==4.1.0
django-otp==1.6.3
geoip2==5.1.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
whitenoise==6.11.0
PyYAML==6.0.3
inflection==0.5.1
packaging==25.0
python-dateutil==2.9.0.post0
pytz==2025.2
qrcode==8.2
requests==2.32.5
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
maxminddb==2.8.2
